+++ tests/testthat/test_ParamSet.R
@@ -239,15 +239,15 @@ test_that("ParamSet$default", {
   expect_equal(ps$default, list(x = NULL))
 })
 
-test_that("is_number / is_categ / is_numeric / is_categoric", {
+test_that("is_number / is_categ / all_numeric / all_categoric", {
   expect_equal(th_paramset_full()$is_number,
     c(th_param_int = TRUE, th_param_dbl = TRUE, th_param_fct = FALSE, th_param_lgl = FALSE))
   expect_equal(th_paramset_full()$is_categ,
     c(th_param_int = FALSE, th_param_dbl = FALSE, th_param_fct = TRUE, th_param_lgl = TRUE))
-  expect_equal(th_paramset_numeric()$is_numeric, TRUE)
-  expect_equal(th_paramset_full()$is_numeric, FALSE)
-  expect_equal(th_paramset_categorical()$is_categorical, TRUE)
-  expect_equal(th_paramset_full()$is_categorical, FALSE)
+  expect_equal(th_paramset_numeric()$all_numeric, TRUE)
+  expect_equal(th_paramset_full()$all_numeric, FALSE)
+  expect_equal(th_paramset_categorical()$all_categorical, TRUE)
+  expect_equal(th_paramset_full()$all_categorical, FALSE)
 })
 
 test_that("ParamSet$ids", {