+++ NEWS.md
@@ -4,7 +4,8 @@
   Required parameter are not checked anymore when new values are added to the
   parameter set.
 * `ParamSet$check_dt()` accepts `data.frame`s.
-
+* Rename `is_numeric` and `is_categorical` to `all_numeric` and
+  `all_categorical`.
 
 # paradox 0.6.0